Reverse order of hint characters
This means that hints which are close together start with the same
prefix, which is usually on the home row.

Hints are therefore more densely placed when you choose the second
character, but I think it might be quicker to choose them, as you can
quickly narrow down the approximate area.

Feedback wanted!